Extract common methods
The Advocate and Advocate Harder cleaners both clearn cracked claim details
in the same way. Extracted to ClearCrackedDetails to avoid duplication.
