+++ app/services/claim_cleaner_service/clear_cracked_details.rb
@@ -1,0 +1,16 @@
+class ClaimCleanerService
+  module ClearCrackedDetails
+    private
+
+    def clear_inapplicable_fields
+      clear_cracked_details if case_type.present? && !requires_cracked_dates?
+    end
+
+    def clear_cracked_details
+      claim.trial_fixed_notice_at = nil if trial_fixed_notice_at
+      claim.trial_fixed_at = nil if trial_fixed_at
+      claim.trial_cracked_at = nil if trial_cracked_at
+      claim.trial_cracked_at_third = nil if trial_cracked_at_third
+    end
+  end
+end
